Fix false positives in unused_declaration
Avoid marking protocol definitions as unused when they have a default
implementation.

Add more tests.
